+++ etomica.apps/etomica/normalmode/SimTargetUmbrella.java
@@ -204,7 +204,7 @@ public class SimTargetUmbrella extends Simulation {
         
         IAction outputAction = new IAction(){
         	public void actionPerformed(){
-        		int idStep = sim.integrator.getStepCount();
+        		long idStep = sim.integrator.getStepCount();
 		        /*
 		         * Histogram
 		         */
