+++ etomica.apps/etomica/normalmode/SimUmbrella.java
@@ -264,7 +264,7 @@ public class SimUmbrella extends Simulation {
         
         IAction outputAction = new IAction(){
         	public void actionPerformed(){
-        		int idStep = sim.integrator.getStepCount();
+        		long idStep = sim.integrator.getStepCount();
 		        /*
 		         * Histogram
 		         */
